Revert "Nothing work: Send welcome message to newly created user"
This reverts commit bca80bdb17a801e67f91207c91f70c7769bb43b4, reversing
changes made to 1798cb66bfd71d0440a9cc8df0f32344cc73a462.
