+++ endpoints/users/model.js
@@ -9,11 +9,9 @@ var gearmanode = require("gearmanode");
 var extend = require("util")._extend;
 var csv = require("to-csv");
 var moment = require("moment");
-var fs = require("fs");
 
 var Province = require ("../../resources/misc/schemas/province");
 var KabKota = require ("../../resources/misc/schemas/kabkota");
-var AdminActivities = require ("../../resources/misc/schemas/adminActivities");
 
 var ResourceUser = require ("../../resources/user");
 var Model = ResourceUser.schemas;
@@ -29,11 +27,6 @@ var ServerModel = ResourceServer.schemas;
 var policy = require("../../policy");
 var UserEnums = ResourceUser.enums(policy);
 var UserStates = UserEnums.States;
-
-// Mailer
-const Mailer = require(__dirname + '/../../scripts/mailer');
-const mailerTemplate = fs.readFileSync(__dirname + '/../../scripts/templates/template_welcomeMessage.txt').toString();
-const mailerConfig = JSON.parse(fs.readFileSync(__dirname + '/../../scripts/config.json'));
 
 var Session
 try{
@@ -569,33 +562,13 @@ User.prototype.create = function (ctx, options, cb) {
         var client = gearmanode.client({servers: self.options.gearmand});
         var job = client.submitJob("createUser", "");
         job.on("complete", function() {
-          // The new user has been created, send welcome message
-          DomainModel.Domain.findOne({_id:data.domain}, function(err, result) {
-            const mailer = new Mailer();
-            data.primaryEmailAddress = data.username+"@"+result.name;
-            data.name = data.profile.name;
-            mailer.sendMail(mailerTemplate, mailerConfig.subjects.welcomeMessage, data.primaryEmailAddress, data)
-          });
           console.log("RESULT: " + job.response);
           cb(null, JSON.parse(job.response));
           client.close();
         });
-        if (ctx.session.user.domain.name === policy.mainDomain) {
-          return cb(err, object);
-        }
-        AdminActivities.create({
-          timestamp : new Date(),
-          localDomain : ctx.session.user.domain.name,
-          by : ctx.session.user.username + '@' + ctx.session.user.domain.name,
-          activity : 'createUser',
-          data : data
-        }, function(err, result){
-          if (err) {
-            console.log(err);
-            return cb(boom.badRequest (err.message));
-          }
-          return cb (null, object);
-        })
+
+
+        return cb (null, object);
       });
     })();
   }
@@ -604,6 +577,7 @@ User.prototype.create = function (ctx, options, cb) {
 }
 
 User.prototype.update = function (ctx, options, cb) {
+
   var self = this;
   var body = options.body;
   var id = ctx.params.id;
@@ -628,7 +602,7 @@ User.prototype.update = function (ctx, options, cb) {
     if (err) {
       return cb (err);
     }
-    var userState = data.state
+
     if (!data) {
       // boom
     }
@@ -637,6 +611,7 @@ User.prototype.update = function (ctx, options, cb) {
         return cb(err);
       }
       data.save(function (err, user){
+        
         if (err) {
           console.log(err);
           return cb(boom.badRequest (err.message));
@@ -685,50 +660,22 @@ User.prototype.update = function (ctx, options, cb) {
           });
         }
         
-        var reply = function() {
-          if (body.alias) {
-            Model.User.update({_id:data._id}, { $set: {"alias" : body.alias}}, function(){
-              DomainModel.Domain.findOne({_id:data.domain}, function(err, result) {
-                newJob(body.source,data.username+"@"+result.name);
-                return cb (null, object);
-              });
-            });
-          } else if (data.alias && !body.alias) {
-            Model.User.update({_id:data._id}, { $unset: {"alias" : data.alias}}, function(){
-              newJob(data.alias,false);
+        if (body.alias) {
+          Model.User.update({_id:data._id}, { $set: {"alias" : body.alias}}, function(){
+            DomainModel.Domain.findOne({_id:data.domain}, function(err, result) {
+              newJob(body.source,data.username+"@"+result.name);
               return cb (null, object);
             });
-          } else {
+          });
+        } else if (data.alias && !body.alias) {
+          Model.User.update({_id:data._id}, { $unset: {"alias" : data.alias}}, function(){
+            newJob(data.alias,false);
             return cb (null, object);
-          }
-        }
-        if (ctx.session.user.domain.name === policy.mainDomain) {
-          return reply();
-        }
-        if (userState != data.state) {
-          var activity;
-          if (data.state === 'active') {
-            activity = 'activateUser';
-          }
-          if (data.state === 'inactive') {
-            activity = 'deactivateUser';
-          }
-          AdminActivities.create({
-            timestamp : new Date(),
-            localDomain : ctx.session.user.domain.name,
-            by : ctx.session.user.username + '@' + ctx.session.user.domain.name,
-            activity : activity,
-            data : data
-          }, function(err, result){
-            if (err) {
-              console.log(err);
-              return cb(boom.badRequest (err.message));
-            }
-            return reply();
-          })
+          });
         } else {
-          reply();
+          return cb (null, object);
         }
+  
       });
     }
 
@@ -791,45 +738,22 @@ User.prototype.activate = function (ctx, options, cb) {
 
 User.prototype.remove = function (ctx, options, cb){
   if (ctx.params.id) {
-    Model.User.findOne({_id : ctx.params.id}, function(err, data){
+    Model.User.remove({_id : ctx.params.id}, function(err){
       if (err) return cb (err);
-      Model.User.remove({_id : ctx.params.id}, function(err){
-        if (err) return cb (err);
-        reply(null, {object : "user", _id : ctx.params._id}, data);
-      })
+      cb (null, {object : "user", _id : ctx.params._id})
     });
-  } else {
+  }
+  else {
+
     if (Array.isArray(ctx.query.ids)) {
-      Model.User.find({_id : { $in : ctx.query.ids}}, function(err, data){
+      Model.User.remove({ _id: { $in : ctx.query.ids} }, function(err){
         if (err) return cb (err);
-        Model.User.remove({ _id: { $in : ctx.query.ids} }, function(err){
-          if (err) return cb (err);
-          reply(null, {object : "user", data : ctx.query.ids}, data)
-        })
+        cb (null, {object : "user", data : ctx.query.ids})
       });
 
     } else {
       return cb (boom.badRequest("invalid arguments"));
     }
-  }
-  var reply = function(err, obj, data) {
-    // This err is always null
-    if (ctx.session.user.domain.name === policy.mainDomain) {
-      return cb(err, obj);
-    }
-    AdminActivities.create({
-      timestamp : new Date(),
-      localDomain : ctx.session.user.domain.name,
-      by : ctx.session.user.username + '@' + ctx.session.user.domain.name,
-      activity : 'removeUser',
-      data : data
-    }, function(err, result){
-      if (err) {
-        console.log(err);
-        return cb(boom.badRequest (err.message));
-      }
-      cb(err, obj);
-    })
   }
 }
