+++ policy.js
@@ -1,9 +1,6 @@
 var compose = require("koa-compose");
 var kasta = require("kasta");
 
-// Other than this will be labeled as local domain
-var mainDomain = 'domain.com';
-
 // configurable list of avalilable roles and access levels
 var config = {
 
@@ -82,12 +79,9 @@ var policy = {
       level: acl.accessLevels.user
     }
   },
-  
-  // Main domain
-  mainDomain: mainDomain,
 
   // APP related policies
-  app: {},
+  app: {}
 };
 
 module.exports = policy;
