+++ scripts/template.txt
@@ -1,8 +1,5 @@
-<p>
 Yth. __name__
-</p>
-<p>
+
 Anda tidak menggunakan akun Anda (__primaryEmailAddress__) selama __inactiveMonths__ bulan.
-</p>
